Remove LuckPerms integration:
LuckPerms loads its player data during the PlayerJoinEvent, and updates player
context on PlayerChangedWorldEvent. Both of these happen too late in the event
for CO to get the needed meta.

There were a couple of hacks in this removed code that held together decently,
but I found that the hack for getting information on player join does not load
any data from the player's groups. I'll revisit this integration later.